# Add base cost to consume_wasm_execution_units host function