--- a/radix-engine-tests/assets/metering/cuttlefish/cost_crypto_utils.csv
+++ b/radix-engine-tests/assets/metering/cuttlefish/cost_crypto_utils.csv
@@ -1,18 +1,18 @@
-Total Cost (XRD)                                                           ,            1.04183166606,    100.0%
-- Execution Cost (XRD)                                                     ,                0.6651783,     63.8%
+Total Cost (XRD)                                                           ,            1.05047606606,    100.0%
+- Execution Cost (XRD)                                                     ,                0.6738227,     64.1%
 - Finalization Cost (XRD)                                                  ,                0.0102517,      1.0%
-- Storage Cost (XRD)                                                       ,            0.36640166606,     35.2%
+- Storage Cost (XRD)                                                       ,            0.36640166606,     34.9%
 - Tipping Cost (XRD)                                                       ,                        0,      0.0%
 - Royalty Cost (XRD)                                                       ,                        0,      0.0%
-Execution Cost Breakdown                                                   ,                 13303566,    100.0%
+Execution Cost Breakdown                                                   ,                 13476454,    100.0%
 - AfterInvoke                                                              ,                      518,      0.0%
 - AllocateNodeId                                                           ,                     1455,      0.0%
 - BeforeInvoke                                                             ,                     7778,      0.1%
 - Blake2b256Hash                                                           ,                       91,      0.0%
 - Bls12381G2SignatureAggregate                                             ,                   243986,      1.8%
-- Bls12381V1AggregateVerify                                                ,                  2001824,     15.0%
-- Bls12381V1FastAggregateVerify                                            ,                   632277,      4.8%
-- Bls12381V1Verify                                                         ,                   461378,      3.5%
+- Bls12381V1AggregateVerify                                                ,                  2001824,     14.9%
+- Bls12381V1FastAggregateVerify                                            ,                   632277,      4.7%
+- Bls12381V1Verify                                                         ,                   461378,      3.4%
 - CheckReference                                                           ,                    80024,      0.6%
 - CloseSubstate                                                            ,                    18576,      0.1%
 - CreateNode                                                               ,                    13376,      0.1%
@@ -25,27 +25,27 @@ Execution Cost Breakdown                                                   ,    
 - MarkSubstateAsTransient                                                  ,                       55,      0.0%
 - OpenSubstate::GlobalFungibleResourceManager                              ,                   121872,      0.9%
 - OpenSubstate::GlobalGenericComponent                                     ,                    43690,      0.3%
-- OpenSubstate::GlobalPackage                                              ,                  3497709,     26.3%
+- OpenSubstate::GlobalPackage                                              ,                  3497709,     26.0%
 - OpenSubstate::InternalFungibleVault                                      ,                    90202,      0.7%
 - OpenSubstate::InternalGenericComponent                                   ,                    18682,      0.1%
 - OpenSubstate::InternalKeyValueStore                                      ,                    40536,      0.3%
 - PinNode                                                                  ,                      180,      0.0%
-- PrepareWasmCode                                                          ,                  2282710,     17.2%
+- PrepareWasmCode                                                          ,                  2282710,     16.9%
 - QueryActor                                                               ,                     1000,      0.0%
-- ReadSubstate                                                             ,                  2353589,     17.7%
+- ReadSubstate                                                             ,                  2353589,     17.5%
 - RunNativeCode::Worktop_drop                                              ,                    17918,      0.1%
 - RunNativeCode::get_amount_FungibleVault                                  ,                    14451,      0.1%
 - RunNativeCode::lock_fee                                                  ,                    45243,      0.3%
-- RunWasmCode::CryptoScrypto_blake2b_256_hash                              ,                    10010,      0.1%
-- RunWasmCode::CryptoScrypto_bls12381_g2_signature_aggregate               ,                   362018,      2.7%
-- RunWasmCode::CryptoScrypto_bls12381_v1_aggregate_verify                  ,                   253847,      1.9%
-- RunWasmCode::CryptoScrypto_bls12381_v1_fast_aggregate_verify             ,                   211296,      1.6%
-- RunWasmCode::CryptoScrypto_bls12381_v1_verify                            ,                    61708,      0.5%
-- RunWasmCode::CryptoScrypto_ed25519_verify                                ,                    43794,      0.3%
-- RunWasmCode::CryptoScrypto_keccak256_hash                                ,                    10039,      0.1%
-- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_key_recover                   ,                    44649,      0.3%
-- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_verify                        ,                    55047,      0.4%
-- RunWasmCode::Faucet_lock_fee                                             ,                    24589,      0.2%
+- RunWasmCode::CryptoScrypto_blake2b_256_hash                              ,                    10811,      0.1%
+- RunWasmCode::CryptoScrypto_bls12381_g2_signature_aggregate               ,                   422162,      3.1%
+- RunWasmCode::CryptoScrypto_bls12381_v1_aggregate_verify                  ,                   295892,      2.2%
+- RunWasmCode::CryptoScrypto_bls12381_v1_fast_aggregate_verify             ,                   246322,      1.8%
+- RunWasmCode::CryptoScrypto_bls12381_v1_verify                            ,                    71202,      0.5%
+- RunWasmCode::CryptoScrypto_ed25519_verify                                ,                    50392,      0.4%
+- RunWasmCode::CryptoScrypto_keccak256_hash                                ,                    10801,      0.1%
+- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_key_recover                   ,                    51211,      0.4%
+- RunWasmCode::CryptoScrypto_secp256k1_ecdsa_verify                        ,                    63252,      0.5%
+- RunWasmCode::Faucet_lock_fee                                             ,                    27840,      0.2%
 - Secp256k1EcdsaKeyRecover                                                 ,                    14705,      0.1%
 - Secp256k1EcdsaVerify                                                     ,                    14705,      0.1%
 - SetCallFrameData                                                         ,                      606,      0.0%