--- a/radix-engine-tests/assets/metering/cuttlefish/cost_flash_loan.csv
+++ b/radix-engine-tests/assets/metering/cuttlefish/cost_flash_loan.csv
@@ -1,10 +1,10 @@
-Total Cost (XRD)                                                           ,            0.62267027592,    100.0%
-- Execution Cost (XRD)                                                     ,               0.46223905,     74.2%
+Total Cost (XRD)                                                           ,            0.62384462592,    100.0%
+- Execution Cost (XRD)                                                     ,                0.4634134,     74.3%
 - Finalization Cost (XRD)                                                  ,                0.0322574,      5.2%
-- Storage Cost (XRD)                                                       ,            0.12817382592,     20.6%
+- Storage Cost (XRD)                                                       ,            0.12817382592,     20.5%
 - Tipping Cost (XRD)                                                       ,                        0,      0.0%
 - Royalty Cost (XRD)                                                       ,                        0,      0.0%
-Execution Cost Breakdown                                                   ,                  9244781,    100.0%
+Execution Cost Breakdown                                                   ,                  9268268,    100.0%
 - AfterInvoke                                                              ,                     1884,      0.0%
 - AllocateNodeId                                                           ,                     5141,      0.1%
 - BeforeInvoke                                                             ,                     7694,      0.1%
@@ -21,11 +21,11 @@ Execution Cost Breakdown                                                   ,    
 - OpenSubstate::GlobalFungibleResourceManager                              ,                   259797,      2.8%
 - OpenSubstate::GlobalGenericComponent                                     ,                    47692,      0.5%
 - OpenSubstate::GlobalNonFungibleResourceManager                           ,                   632452,      6.8%
-- OpenSubstate::GlobalPackage                                              ,                  3160161,     34.2%
+- OpenSubstate::GlobalPackage                                              ,                  3160161,     34.1%
 - OpenSubstate::InternalFungibleVault                                      ,                   290575,      3.1%
 - OpenSubstate::InternalGenericComponent                                   ,                   212121,      2.3%
 - PinNode                                                                  ,                      636,      0.0%
-- PrepareWasmCode                                                          ,                   967836,     10.5%
+- PrepareWasmCode                                                          ,                   967836,     10.4%
 - QueryActor                                                               ,                    10500,      0.1%
 - ReadSubstate                                                             ,                  1325876,     14.3%
 - RunNativeCode::AuthZone_pop                                              ,                    33696,      0.4%
@@ -47,7 +47,7 @@ Execution Cost Breakdown                                                   ,    
 - RunNativeCode::lock_fee                                                  ,                   116047,      1.3%
 - RunNativeCode::mint_ruid_NonFungibleResourceManager                      ,                    93979,      1.0%
 - RunNativeCode::on_drop_FungibleProof                                     ,                    28382,      0.3%
-- RunNativeCode::on_move_FungibleProof                                     ,                    32360,      0.4%
+- RunNativeCode::on_move_FungibleProof                                     ,                    32360,      0.3%
 - RunNativeCode::put_FungibleBucket                                        ,                    16165,      0.2%
 - RunNativeCode::put_FungibleVault                                         ,                    49108,      0.5%
 - RunNativeCode::take_FungibleBucket                                       ,                    19855,      0.2%
@@ -55,8 +55,8 @@ Execution Cost Breakdown                                                   ,    
 - RunNativeCode::try_deposit_batch_or_abort                                ,                   121257,      1.3%
 - RunNativeCode::unlock_amount_FungibleVault                               ,                    46544,      0.5%
 - RunNativeCode::withdraw                                                  ,                    57851,      0.6%
-- RunWasmCode::BasicFlashLoan_repay_loan                                   ,                    75849,      0.8%
-- RunWasmCode::BasicFlashLoan_take_loan                                    ,                    69367,      0.8%
+- RunWasmCode::BasicFlashLoan_repay_loan                                   ,                    88411,      1.0%
+- RunWasmCode::BasicFlashLoan_take_loan                                    ,                    80292,      0.9%
 - SetCallFrameData                                                         ,                      606,      0.0%
 - SwitchStack                                                              ,                     1000,      0.0%
 - ValidateTxPayload                                                        ,                    18800,      0.2%