--- a/radix-transaction-scenarios/generated-examples/cuttlefish/access-controller-v2/costings/001--access-controller-v2-instantiate.txt
+++ b/radix-transaction-scenarios/generated-examples/cuttlefish/access-controller-v2/costings/001--access-controller-v2-instantiate.txt
@@ -1,10 +1,10 @@
-Total Cost (XRD)                                                           ,            0.94481573742,    100.0%
-- Execution Cost (XRD)                                                     ,               0.32974775,     34.9%
+Total Cost (XRD)                                                           ,            0.94522588742,    100.0%
+- Execution Cost (XRD)                                                     ,                0.3301579,     34.9%
 - Finalization Cost (XRD)                                                  ,                0.1960235,     20.7%
-- Storage Cost (XRD)                                                       ,            0.41904448742,     44.4%
+- Storage Cost (XRD)                                                       ,            0.41904448742,     44.3%
 - Tipping Cost (XRD)                                                       ,                        0,      0.0%
 - Royalty Cost (XRD)                                                       ,                        0,      0.0%
-Execution Cost Breakdown                                                   ,                  6594955,    100.0%
+Execution Cost Breakdown                                                   ,                  6603158,    100.0%
 - AfterInvoke                                                              ,                      846,      0.0%
 - AllocateNodeId                                                           ,                     3201,      0.0%
 - BeforeInvoke                                                             ,                     5962,      0.1%
@@ -44,8 +44,8 @@ Execution Cost Breakdown                                                   ,    
 - RunNativeCode::lock_fee                                                  ,                    45243,      0.7%
 - RunNativeCode::put_FungibleVault                                         ,                    24554,      0.4%
 - RunNativeCode::take_FungibleVault                                        ,                    42457,      0.6%
-- RunWasmCode::Faucet_free                                                 ,                    34815,      0.5%
-- RunWasmCode::Faucet_lock_fee                                             ,                    24589,      0.4%
+- RunWasmCode::Faucet_free                                                 ,                    39767,      0.6%
+- RunWasmCode::Faucet_lock_fee                                             ,                    27840,      0.4%
 - SetCallFrameData                                                         ,                      606,      0.0%
 - SetSubstate                                                              ,                      944,      0.0%
 - SwitchStack                                                              ,                     1000,      0.0%