--- a/radix-transaction-scenarios/generated-examples/cuttlefish/access-controller-v2/receipts/003--access-controller-v2-lock-fee-and-recover.txt
+++ b/radix-transaction-scenarios/generated-examples/cuttlefish/access-controller-v2/receipts/003--access-controller-v2-lock-fee-and-recover.txt
@@ -172,7 +172,7 @@ STATE UPDATES: 5 entities
        Value: UNLOCKED ConsensusManagerValidatorRewardsFieldPayload::V1(
          ValidatorRewardsSubstate {
            proposer_rewards: {
-             0u8 => Decimal("0.426474117095"),
+             0u8 => Decimal("0.426679192095"),
            },
            rewards_vault: Vault(Own("internal_vault_sim1tpsesv77qvw782kknjks9g3x2msg8cc8ldshk28pkf6m6lkhun3sel")),
          },
@@ -270,7 +270,7 @@ STATE UPDATES: 5 entities
   └─ Partition(64): 1 change
     └─ Set: Field(0)
        Value: UNLOCKED FungibleVaultBalanceFieldPayload::V1(
-         LiquidFungibleResource(Decimal("0.85294823419")),
+         LiquidFungibleResource(Decimal("0.85335838419")),
        )
 
 OUTPUTS: 3