## Summary
Added WASM execution unit base to assure simple WASM calls are not being undercharged.

## Motivation

Test | Base | PR | %
-- | -- | -- | --
costing::scrypto_malloc | 592.2±0.84ms | 504.3±2.87ms | -14.84%
costing::scrypto_sbor_decode | 569.5±1.32ms | 500.3±2.76ms | -12.15%
costing::scrypto_sha256 | 392.4±1.20ms | 367.5±1.57ms | -6.35%
costing::spin_loop_v1 | 628.2±5.74ms | 474.9±2.19ms | -24.40%
costing::spin_loop_v2 | 2.7±0.00s | 959.4±0.90ms | -64.47%

Based on the benchmarks on the `base`, the `spin_loop_v2` does not finish within 1 second, despite consuming around 100 million execution cost units. 

It's made possible by iterating an empty code block (except a `br` instruction), where `consume_wasm_execution_units` host function becomes the dominant factor. 

This PR fixes it by introducing a base cost to account for the execution of the host function itself.